move filldagger and mergeAllNodes to utils
